Guard against handling errors multiple times.
There are some cases where a connection can raise the 'error' event
multiple times. For example, if you can connect to the SQL Server
Browser Service but the instance you are looking for isn't configured
for TCP/IP, then you'll get a "Port for ... not found" error from
InstanceLookup followed by a "Failed to connect to ... in ... ms" from
the connection itself.

This can lead to making more and more attempts to connect over time as
we schedule multiple retries for each connection attempt.